+++ src/frontend/pages/components/gamelistpage/GameListItemInfo.tsx
